requirements.txt
langchain
langchain-core
langchain-community
langgraph
groq
python-dotenv
pydantic
tqdm
faiss-cpu
sentence-transformers